prova

+++ gridpack_generation_EFT.sh
@@ -31,7 +31,11 @@ make_tarball () {
         XZ_OPT="--lzma2=preset=9,dict=512MiB"
     fi
 
-    mkdir InputCards
+    if [ -d InputCards ]; then 
+      rm -rf InputCards
+    fi
+
+    mkdir -p InputCards
     cp $CARDSDIR/${name}*.* InputCards
 
     EXTRA_TAR_ARGS=""
@@ -186,8 +190,6 @@ make_gridpack () {
       tar xzf $SMEFT
       rm $SMEFT
       cd ..;
-
-
       #############################################
       #Apply any necessary patches on top of official release
       #############################################
@@ -382,7 +384,8 @@ make_gridpack () {
           if [ "${BASH_SOURCE[0]}" != "${0}" ]; then return 0; else exit 0; fi
       fi
     
-    elif [ "${jobstep}" = "INTEGRATE" ] || [ "${jobstep}" = "ALL" ]; then  
+    #elif [ "${jobstep}" = "INTEGRATE" ] || [ "${jobstep}" = "ALL" ]; then  
+    elif [ "${jobstep}" = "INTEGRATE" ] || [ "${jobstep}" = "ALL" || [ "${jobstep}" = "REWEIGHT" ]; then
       echo "Reusing existing directory assuming generated code already exists"
       echo "WARNING: If you changed the process card you need to clean the folder and run from scratch"
     
@@ -432,16 +435,27 @@ make_gridpack () {
     
     fi  
     
-    if [ -d gridpack ]; then
-      rm -rf gridpack
-    fi
+    if [ "${jobstep}" = "INTEGRATE" ]; then 
+      # If integrate step then clean up folder
+      # prom previous (unsuccessful ? ) runs
+      # and start from scratch
+
+      # if not on integrate step we want to keep these 
+      # folders (e.g. reweight) or at least some of them 
+      # because they store all the integrate information 
     
-    if [ -d processtmp ]; then
-      rm -rf processtmp
-    fi
-    
-    if [ -d process ]; then
-      rm -rf process
+      if [ -d gridpack ]; then
+        rm -rf gridpack
+      fi
+      
+      if [ -d processtmp ]; then
+        rm -rf processtmp
+      fi
+      
+      if [ -d process ]; then
+        rm -rf process
+      fi
+
     fi
     
     if [ ! -d ${name} ]; then
@@ -511,6 +525,9 @@ make_gridpack () {
       cp $CARDSDIR/${name}_param_card.dat ./Cards/param_card.dat
     fi
      
+    if [ "${jobstep}" = "INTEGRATE" ] || [ "${jobstep}" = "ALL" ]; then
+      echo "---> SONO QUI DENTRO"
+      echo $jobstep
     if [ "$isnlo" -gt "0" ]; then
     #NLO mode  
       #######################
@@ -568,7 +585,7 @@ make_gridpack () {
       if [ -e $CARDSDIR/${name}_externaltarball.dat ]; then
         mv $WORKDIR/header_for_madspin.txt . 
       fi
-      
+
     else
       #LO mode
       #######################
@@ -586,6 +603,7 @@ make_gridpack () {
       echo "done" >> makegrid.dat
     
     #   set +e
+      sed -i.bak '/FFLAGS= -O -w -fbounds-check -fPIC/s/$/ -mcmodel=medium/' Source/make_opts 
       cat makegrid.dat | ./bin/generate_events pilotrun
       echo "finished pilot run"
     
@@ -602,19 +620,19 @@ make_gridpack () {
       echo "cleaning temporary output"
       mv $WORKDIR/processtmp/pilotrun_gridpack.tar.gz $WORKDIR/
       mv $WORKDIR/processtmp/Events/pilotrun/unweighted_events.lhe.gz $WORKDIR/
-      rm -rf processtmp
+      #rm -rf processtmp
       mkdir process
       cd process
       echo "unpacking temporary gridpack"
       tar -xzf $WORKDIR/pilotrun_gridpack.tar.gz
       echo "cleaning temporary gridpack"
-      rm $WORKDIR/pilotrun_gridpack.tar.gz
+      #rm $WORKDIR/pilotrun_gridpack.tar.gz
 
       # precompile reweighting if necessary
       if [ -e $CARDSDIR/${name}_reweight_card.dat ]; then
           echo "preparing reweighting step"
           prepare_reweight $isnlo $WORKDIR $scram_arch $CARDSDIR/${name}_reweight_card.dat
-	  extract_width $isnlo $WORKDIR $CARDSDIR ${name}
+	      extract_width $isnlo $WORKDIR $CARDSDIR ${name}
       fi
       
       #prepare madspin grids if necessary
@@ -635,15 +653,98 @@ make_gridpack () {
       
       cd $WORKDIR
       
-      mkdir gridpack
-      mv process gridpack/process
-      cp -a $MGBASEDIRORIG/ gridpack/mgbasedir
+        mkdir -p gridpack
+
+        if [ -d gridpack/process ]; then 
+          rm -rf gridpack/process
+        fi 
+
+        mv process gridpack/process
+        cp -a $MGBASEDIRORIG/ gridpack/mgbasedir
     
       cd gridpack
       
       cp $PRODHOME/runcmsgrid_LO.sh ./runcmsgrid.sh
     fi
+    fi
     
+    # # # # # # # # # # # # #
+    # Here starts reweight  #
+    # # # # # # # # # # # # #
+
+    if [ "${jobstep}" = "REWEIGHT" ]; then
+
+      # exit on first error
+      set +e 
+      cd $WORKDIR
+
+
+      echo "----> starting reweighting!! "
+
+      echo "cleaning temporary output"
+      echo $PWD
+      
+      # It should be here after codegen
+      if [ ! -d process ]; then 
+        # If not that could mean that integrate moved it under gridpack (line  624 for LO)
+        if [ -d gridpack/process ]; then 
+          # bring it back :)
+          mv gridpack/process .
+        else
+          # otherwise create it 
+          mv $WORKDIR/processtmp/pilotrun_gridpack.tar.gz $WORKDIR/
+          mkdir -p process
+          cd process
+          echo "unpacking temporary gridpack"
+          tar -xzf $WORKDIR/pilotrun_gridpack.tar.gz
+        fi
+      fi
+
+      if [ -d process/madevent/rwgt ]; then 
+        rm -rf process/madevent/rwgt
+      fi
+
+      if [ ! -f $WORKDIR/unweighted_events.lhe.gz ]; then 
+        mv $WORKDIR/processtmp/pilotrun/unweighted_events.lhe.gz $WORKDIR/
+      fi
+
+      # precompile reweighting if necessary
+      if [ -e $CARDSDIR/${name}_reweight_card.dat ]; then
+          echo "preparing reweighting step"
+          prepare_reweight $isnlo $WORKDIR $scram_arch $CARDSDIR/${name}_reweight_card.dat
+          extract_width $isnlo $WORKDIR $CARDSDIR ${name}
+      fi
+      
+      #prepare madspin grids if necessary
+      if [ -e $CARDSDIR/${name}_madspin_card.dat ]; then
+        echo "import $WORKDIR/unweighted_events.lhe.gz" > madspinrun.dat
+        cat $CARDSDIR/${name}_madspin_card.dat >> madspinrun.dat
+        $WORKDIR/$MGBASEDIRORIG/MadSpin/madspin madspinrun.dat 
+        rm madspinrun.dat
+        rm -rf tmp*
+      fi
+    
+      echo "preparing final gridpack"
+      
+      #set to single core mode
+      echo "mg5_path = ../../mgbasedir" >> ./madevent/Cards/me5_configuration.txt
+      echo "cluster_temp_path = None" >> ./madevent/Cards/me5_configuration.txt
+      echo "run_mode = 0" >> ./madevent/Cards/me5_configuration.txt  
+      
+      cd $WORKDIR
+      
+      if [ -d gridpack ]; then
+        mkdir -p gridpack
+        mv process gridpack/process
+        cp -a $MGBASEDIRORIG/ gridpack/mgbasedir
+      fi
+    
+      cd gridpack
+      
+      cp $PRODHOME/runcmsgrid_LO.sh ./runcmsgrid.sh
+
+    fi
+
     sed -i s/SCRAM_ARCH_VERSION_REPLACE/${scram_arch}/g runcmsgrid.sh
     sed -i s/CMSSW_VERSION_REPLACE/${cmssw_version}/g runcmsgrid.sh
     
@@ -697,7 +798,7 @@ queue=${3}
 
 # processing options
 jobstep=${4}
-
+echo $jobstep
 # sync default cmssw with the current OS 
 export SYSTEM_RELEASE=`cat /etc/redhat-release`
 
@@ -739,7 +840,7 @@ fi
 helpers_dir=${PRODHOME%genproductions*}/genproductions/Utilities
 helpers_file=${helpers_dir}/gridpack_helpers.sh
 if [ ! -f "$helpers_file" ]; then
-  if ! [ -x "$(command -v git)" ]; then
+  if  [ -x "$(command -v git)" ]; then
     helpers_dir=${PRODHOME}/Utilities
   else
     helpers_dir=$(git rev-parse --show-toplevel)/bin/MadGraph5_aMCatNLO/Utilities
@@ -776,7 +877,7 @@ if [ -z ${jobstep} ]; then
 fi
 
 #Check values of jobstep:
-if [ "${jobstep}" == "ALL" ] || [ "${jobstep}" == "CODEGEN" ] || [ "${jobstep}" == "INTEGRATE" ] || [ "${jobstep}" == "MADSPIN" ]; then
+if [ "${jobstep}" == "ALL" ] || [ "${jobstep}" == "CODEGEN" ] || [ "${jobstep}" == "INTEGRATE" ] || [ "${jobstep}" == "REWEIGHT" ] || [ "${jobstep}" == "MADSPIN" ]; then
     echo "Running gridpack generation step ${jobstep}"
 else
     echo "No Valid Job Step specified, exiting "
